🔧 Move MatchApiError class to its own file and update imports in match API
